requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# ─────────────────────────────────────────────
# BACKEND
# ─────────────────────────────────────────────
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.12.5
python-dotenv==1.2.1
httpx==0.28.1
requests==2.32.3
python-multipart

# ─────────────────────────────────────────────
# ML / EMBEDDINGS
# ─────────────────────────────────────────────
torch==2.7.0+cpu
sentence-transformers==3.4.1
transformers==4.51.3
tokenizers==0.21.0
safetensors
# huggingface_hub intentionally unpinned — let pip resolve between transformers + sentence-transformers
numpy==1.26.4
scipy
scikit-learn

# ─────────────────────────────────────────────
# RAG PIPELINE
# ─────────────────────────────────────────────
langgraph==0.4.1
langchain-core==0.3.55
qdrant-client==1.13.3
rank-bm25==0.2.2

# ─────────────────────────────────────────────
# LLM PROVIDERS
# ─────────────────────────────────────────────
openai==1.77.0
groq==1.0.0
google-generativeai==0.8.3
google-genai

# ─────────────────────────────────────────────
# GOOGLE AUTH & DRIVE API
# ─────────────────────────────────────────────
google-api-python-client==2.190.0
google-auth==2.38.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0

# ─────────────────────────────────────────────
# PDF & DOCUMENT PARSING
# ─────────────────────────────────────────────
pdfplumber==0.11.9
pdfminer.six==20251230
PyMuPDF==1.26.7
pdf2image==1.17.0
pytesseract==0.3.13
python-docx==1.2.0
pillow>=9.1,<12

# ─────────────────────────────────────────────
# DATA PROCESSING
# ─────────────────────────────────────────────
pandas==2.2.2
pyarrow==23.0.1

# ─────────────────────────────────────────────
# DATABASE / STORAGE
# ─────────────────────────────────────────────
supabase

# ─────────────────────────────────────────────
# UTILS
# ─────────────────────────────────────────────
loguru==0.7.3
tenacity>=8.0.0
tqdm==4.67.1
PyYAML==6.0.2
python-dateutil==2.9.0.post0
pytz==2025.1
tzdata==2025.1
certifi==2025.1.31